# backend/src/main/java/com/product/template/controller/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.product.template.controller;

import com.product.template.dto.ErrorResponse;
import com.product.template.dto.LoginRequest;
import com.product.template.dto.RefreshTokenRequest;
import com.product.template.dto.SignupRequest;
import com.product.template.dto.TokenResponse;
import com.product.template.dto.UserResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.12.0")
@Validated
@Tag(name = "Auth", description = "認証関連API")
public interface AuthApi {

    /**
     * POST /api/auth/login : ログイン
     * メールアドレスとパスワードで認証し、アクセストークンとリフレッシュトークンを取得する
     *
     * @param loginRequest  (required)
     * @return 認証成功 (status code 200)
     *         or バリデーションエラー (status code 400)
     *         or 認証失敗（メールアドレスまたはパスワードが不正） (status code 401)
     */
    @Operation(
        operationId = "login",
        summary = "ログイン",
        description = "メールアドレスとパスワードで認証し、アクセストークンとリフレッシュトークンを取得する",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "認証成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TokenResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "バリデーションエラー", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "認証失敗（メールアドレスまたはパスワードが不正）", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/login",
        produces = { "application/json" },
        consumes = "application/json"
    )
    
    ResponseEntity<TokenResponse> login(
        @Parameter(name = "LoginRequest", description = "", required = true) @Valid @RequestBody LoginRequest loginRequest
    );


    /**
     * POST /api/auth/logout : ログアウト
     * 現在のユーザーのリフレッシュトークンを無効化する
     *
     * @return ログアウト成功 (status code 204)
     *         or 未認証 (status code 401)
     */
    @Operation(
        operationId = "logout",
        summary = "ログアウト",
        description = "現在のユーザーのリフレッシュトークンを無効化する",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "204", description = "ログアウト成功"),
            @ApiResponse(responseCode = "401", description = "未認証", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/logout",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> logout(
        
    );


    /**
     * POST /api/auth/refresh : トークンリフレッシュ
     * リフレッシュトークンを使用して新しいアクセストークンを取得する
     *
     * @param refreshTokenRequest  (required)
     * @return トークンリフレッシュ成功 (status code 200)
     *         or バリデーションエラー (status code 400)
     *         or リフレッシュトークンが無効または期限切れ (status code 401)
     */
    @Operation(
        operationId = "refreshToken",
        summary = "トークンリフレッシュ",
        description = "リフレッシュトークンを使用して新しいアクセストークンを取得する",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "トークンリフレッシュ成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TokenResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "バリデーションエラー", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "リフレッシュトークンが無効または期限切れ", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/refresh",
        produces = { "application/json" },
        consumes = "application/json"
    )
    
    ResponseEntity<TokenResponse> refreshToken(
        @Parameter(name = "RefreshTokenRequest", description = "", required = true) @Valid @RequestBody RefreshTokenRequest refreshTokenRequest
    );


    /**
     * POST /api/auth/signup : ユーザー登録
     * 新規ユーザーを登録する
     *
     * @param signupRequest  (required)
     * @return ユーザー登録成功 (status code 201)
     *         or バリデーションエラー (status code 400)
     *         or メールアドレスが既に登録済み (status code 409)
     */
    @Operation(
        operationId = "signup",
        summary = "ユーザー登録",
        description = "新規ユーザーを登録する",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "201", description = "ユーザー登録成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "バリデーションエラー", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "メールアドレスが既に登録済み", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/signup",
        produces = { "application/json" },
        consumes = "application/json"
    )
    
    ResponseEntity<UserResponse> signup(
        @Parameter(name = "SignupRequest", description = "", required = true) @Valid @RequestBody SignupRequest signupRequest
    );

}